Do inference in integration test (#39)
By calling the inference script as part of the integration test, we get
some coverage of the model checkpointing/loading. This is also a better
reflection of our standard training procedure which includes a call to
the inference script after training completes. A downside of this PR is
that the integration test is slower: about 30s now, counting both model
architectures.

Adding inference to the integration test revealed a bug in the code in
which training was exiting before the last epoch is checkpointed. This
bug is fixed.

Resolves #31 and resolves #32