Delete unused features in train.py and inference.py (#44)
There are various features which we are not using. And if we want to use
a similar feature in the future (e.g. adding new 'fixed' inputs like
orography or something else) we would want to implement in a more
general way. So deleting these features will make it easier to
understand and refactor the existing code. There are also some changes
to the formatting (mostly trying to reduce line lengths so that code is
more easily readable).

This PR will be easier to review commit-by-commit.