Pin dependencies installed in docker image build (#46)
Currently the packages installed in our docker image do not have their
versions pinned or defined anywhere. This can lead to code breaking
unexpectedly (for example #47). Also there is no clear definition of the
requirements for the `fme` package which makes development and CI
tricky. This PR improves the situation.

Note, the strategy of compiling a `constraints.txt` file using
`pip-compile` (as in fv3net) was attempted but abandoned. This was
mostly because we do not want to install a new version of pytorch in the
docker image. If we installed requirements using `pip install -c
constraints.txt -r requirements.txt`, a version of pytorch would get
installed which is different from the version already in the docker
image. Therefore, I just used `pip freeze` to pin the versions in a new
`requirements_except_torch.txt.` Maybe this file should actually be in
the `docker` directory, or the top-level of the `full-model` repo.

This leads to a much simpler Dockerfile, and should make it easier to
create a consistent environment e.g. for testing in CI or local
development.

Resolves #36 and resolves #47