Add GitHub action to run FME unit tests (#51)
Add a GitHub action for running all unit tests that do not require a
GPU. As part of this, I developed some tooling for creating a "fme"
conda environment with our dependencies installed. This may be useful
for development that does not require a GPU and so can be done in local
environments (i.e. on your mac).

I found it difficult to pip install the `dgl` library (a dependency of
the GraphCastNet architecture), and so to avoid requiring a conda
environment for our GitHub CI action, I chose to comment out the import
of this architecture. The error I get otherwise is shown in the comment
below.

Resolves #37