Remove wind model and two step training (#80)
The logic implementing two-step-training assumes the time dimension is
folded into the channel dimension, and will be entirely re-written. It
is currently not a usable option as the FV3GFS data loader does not
support it, and when we do add support, we will do this by adding a time
dimension. This PR deletes that code so we can re-implement the
functionality in a more extensible n-step-training way.

This PR also deletes the dead code for loading the wind model.