Add script to compute "vertically resolved" FV3GFS dataset (#77)
Add a script to generate a new zarr dataset that includes the variables
defined in this doc:
https://docs.google.com/document/d/15Nrg4x9j8MTNb31ne8ST1wWCJc45_n8Ov3GUy_IWN6k/edit#heading=h.vgieay8l2nph

You can see a notebook which plots all of the outputs for a single
snapshot here:
https://github.com/ai2cm/explore/blob/master/oliwm/2023-04-16-fme-analysis/2023-05-10-check-vertically-resolved-dataset.ipynb

The resulting dataset is about 238GB (the input is about 2.7TB). Running
this script on my 8-CPU VM takes about 2.5 hours. We could write some
code to parallelize the job across multiple workers, but it doesn't seem
worth the trouble right now. 2.5 hours is not that long given we don't
expect to run this too often.