Enable input-only and output-only variables (#89)
Changes:
- SingleModuleStepper now lives in its own module
- New Normalizer and Packer classes are responsible for normalizing and
packing/unpacking data into tensors
- Data outside of the stepper is now in the form of Mapping[str,
torch.Tensor], which are denormalized coming from the dataset. The
stepper is responsible for normalizing/packing/unpacking these into
tensors for modules
- Optimization logic is contained in its own logic, so that at inference
time no optimization needs to be configured
- Loaded data now has a time axis, instead of being split into input and
target data. The stepper is responsible for selecting data from the
input or target timestep.
- The stepper now takes in a ModuleBuilder and data_shapes instead of a
Module, so that the ModuleBuilder can supply data requirements to the
data loader, and the data loader can in turn supply data shapes for the
Module initialization, without creating a circular dependency.
- xarray is added as a requirement for IO, and is used to save predicted
timeseries in the inference script

---------

Co-authored-by: Oliver Watt-Meyer <oliverwm@allenai.org>