Use consistent module structure on multi- and single-gpu systems (#126)
Currently, our module structure starts with a top-level
DistributedDataParallel module with a `.module` attribute if we are on a
multi-GPU system, and no top-level wrapper otherwise. This makes it
difficult to save the model on one system in one of these configurations
and load it in another configuration.

This PR modifies the non-distributed behavior to add a dummy wrapper
with a `.module` attribute, so the module structure is always identical
and modules can easily be dumped and loaded between systems.