Use self.num_batches_seen for wandb step instead of epoch (#148)
A recent PR #135 was added which changes the meaning of a wandb step to
"number of batches seen", but when merging main some code was restored
which uses wandb step as number of epochs. This has resulted
in wandb not reporting any epoch-level logs. This PR restores the
behavior of using number of batches seen everywhere, which will
cause the wandb logs to return.