Delete test_fv3gfs_dataset module which has very out-of-date tests (#168)
This module was used to develop an early version of the "FV3GFS" data
loader. This data loader has changed substantially so the tests are no
longer relevant or useful (and are skipped by default anyway). Perhaps
we should add unit tests of the data loader, but for now its best to
clean up by deleting these out of date tests.