Allow standalone inference on more than one sample (#178)
This PR updates the standalone inference so it can be configured and run
on more than one sample.

Functionally there are no other changes, though because of
confusing/hacky ways the APIs were previously used, it should be noted
that aggregators are no longer used to store any netCDF data during
inference. These were previously only used to store ensemble-mean
values, but only on a single value so that it corresponded to raw data.
Because of this previous (and now removed) restriction, no user
functionality has changed. Saving ensemble-mean [time, lat, lon] metrics
is left to a future PR.

Instead, raw data is stored directly by appending to netCDF using
netcdf4-python.

No changes are made to inference during training, which is still
hard-coded to run on a single ensemble member. Updating this is also
left for a future PR.