Add window_time_slice argument to get_data_loader (#186)
We want to enable inference on very long rollout windows. This requires
being able to initialize a series of data loaders which will get
progressive slices of the overall rollout windows. We enable this by
adding a `window_time_slice` argument to `get_data_loader`. If given,
the initialized loader will return only that time slice of each window.