Add config to interpolate prescribed and generated data (#198)
We have gridcells with fractional ocean area coverage. In these cells it
may be more appropriate to interpolate between generated and prescribed
surface temperature values, rather than replacing them completely, to
give values which play better with the spherical harmonic transform.

This PR adds a configuration option allowing this to be done, assuming
the mask variable is 1 where values should be prescribed and 0 where
they should not. It maintains backwards compatibility with the existing
code, and is disabled by default.