Make datasets type safe (#209)
This PR resolves a few issues with our dataset handling:
- As we find more requirements of our datasets, we've been adding new
attributes to DataLoader. This causes typing issues and `type: ignore`
statements in other parts of the code that expect a torch DataLoader, as
this type does not have the extra attributes. This PR moves them to a
new statically-typed class.
- `get_data_loader` has several return values and its call signature is
input-dependent. This PR updates the return value to a static type that
can be accessed by attributes instead of less clear integer indexing.
- It is difficult to detect whether our datasets support different
required features. This PR adds an abstract base class to be used in
each data loader, to ensure it provides the features we need. While this
is still extending the DataLoader type, because the dependence on this
ABC is limited to the data loading code, we only need to reference this
special type within that code and can reference torch DataLoader
throughout the training and inference code.

---------

Co-authored-by: James Duncan <jpdunc23@users.noreply.github.com>