Reduce memory req for inference when using only one time window (#223)
This PR updates EnsembleBatch so that on the last window of rollout, it
doesn't save the initial condition for a non-existent next window. This
allows us to run inference with ensembles too large to fit even the
initial condition in memory, provided the inference rollout is done as a
single time window.