Add dataset-based inference run capability (#216)
This PR adds an optional key `prediction_data` to the `InferenceConfig`
which, when provided, will use that data instead of model data for
evaluation. This enables us to evaluate traditional models (such as
lower-resolution versions of FV3GFS) through the data loaders.

It also adds object-based infrastructure for creating mock data during
tests, as I found the existing routines tightly coupled and hard to
extend. We may want to refactor the existing tests to use this.