Add initial_conditions slice arg to DataLoaderParams (#273)
We would like to be able to train excluding the first set of samples
from our training data, and also run inference starting from
spaced-in-time initial conditions. To enable this, this PR adds a
`window_starts` argument to DataLoaderParams which lets you specify a
slice for the allowed initial conditions indices from the data.