Better handling of missing atmospheric fields in climate data object (#285)
This PR centralizes the handling of missing atmospheric fields into the
climate data object. Specifically, instead of yielding an exception, it
now returns logs a warning and returns None, thus leaving the handling
of this case up to the caller. An additional advantage of this approach
is that the type checker notices if the `None` case is not handled.

Also, adds a unit test for climate data.

Closes #266

---------

Co-authored-by: James Duncan <jpdunc23@users.noreply.github.com>
Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>