Add configurable loss on global mean (#305)
We'd like to directly encourage the model to accurately predict the
evolution of global-mean values. To do this, we add a LossConfig which
lets the user add in a component to the loss computed on global means.

This allows an arbitrary loss (currently only LpLoss) to be configured
for the global-mean value. To do an L1 loss, we would set a config like:

```
stepper:
  loss:
    type: LpLoss
    kwargs:
      p: 2
    global_mean_type: LpLoss
    global_mean_kwargs:
      p: 1
    global_mean_weight: 1.0
```

Changes:
- added LossConfig under SingleModuleStepperConfig.

- [x] Tests added