Allow fine-tuning existing models (#320)
Fine-tuning is often discussed in the FME weather literature and may
prove to be important for us as well. This PR implements this feature by
introducing a pattern for loading certain components off disk, and
creating other ones fresh from a config. For example, loading an
existing stepper from a previous training job, and fine-tuning it using
a different optimizer. One of the consequences of this implementation is
that resuming a fine-tuning job is no different from resuming a regular
job. The "fine-tuning" aspect of the configuration disappears as soon as
the job is run.

Changes:
- Creates a class `ExistingStepperConfig` which corresponds to a
different `stepper` block in the config.


- [x] Tests added