Enable EMA for training and inference (#314)
@salvaRC found that using the EMA of the model resulted in more stable
behavior than using a final model state. We'd like to be able to
evaluate EMA versions of our full model emulator to increase model
stability. This PR adds that functionality.

Changes:
- Copied LitEMA from
https://github.com/ai2cm/full-model/blob/salva/src/models/modules/ema.py
to `fme.core.ema.EMATracker`
- SingleModuleStepper.modules is now `torch.nn.ModuleList` instead of
`List[torch.nn.Module]`
- `ema` is now added to `TrainConfig`, and is enabled by default,
controlling the settings for the EMA version of the model
- `validate_using_ema` is added to `TrainConfig` (default False) and
controls whether validation and inference use the EMA version of the
model
- Training now saves a new file `ema_ckpt.tar` which contains a version
of the stepper with the EMA model weights. The existing checkpoint files
are unchanged.