Generalize training metrics and log the autoregressive loss at each step to WandB. (#349)
We need to track the loss at each step of the autoregressive loss. This
PR achieves that by generalizing `SteppedData` to contain a dictionary
of metrics, not just a single loss.

Changes:
- Replaces `SteppedData.loss` with `SteppedData.metrics: Dict[str,
torch.Tensor]`. Unfortunately, we need the values to be torch tensors so
that they can be reduced when running in distributed mode.
- At each autoregressive step, adds `loss_step_i` for `i in [0,
n_forward_steps]` to metrics.

Resolves #345