Only apply residual scaling to prognostic variables (#347)
Previously we applied the "residual scaling" approach to all variables.
But it doesn't make sense to do it for input-only or output-only
variables. This PR restricts the residual scaling to only apply to
prognostic variables. It also gets rid of the confusing "geometric mean"
rescaling, and uses a simpler approach of making all normalized
prognostic variables have a standard deviation of at least 1 (so they
will be weighted at least as much as the diagnostic variables).

Changes:
- only do residual scaling for prognostic variables
- don't use geometric mean, instead set minimum standard deviation of
residucal-scaled prognostic variables to 1

Resolves #153