Output inference diagnostics as netcdf for paper-ready figures (#333)
We need to make pretty plots for our paper and wandb isn't intended for
this. This PR outputs much of the raw data (e.g. maps) that are sent to
wandb in netcdf form so that it can be slurped up and piped to custom
plotting scripts.

Changes:
- Refactors `time_mean.py` to facilitate code re-use between
`get_dataset` and `get_logs`.
- Passes coords to `zonal_mean.py` so that outputs have the correct
components.
- Adds `get_datasets` method to `InferenceAggregator` so that all the
xarray Datasets can be passed up to the top-level for saving.
- Writes each aggregators' dataset to disk in netcdf format.

[x] Tests added

Resolves #298

---------

Co-authored-by: Oliver Watt-Meyer <oliverwatt@gmail.com>