Implement DynamicHistogram (#362)
We would like to compute histogram metrics during inference within
aggregators. One difficulty with doing this dynamically is that the size
we need to make the bins is not known until all samples are seen. To get
around this, we define a DynamicHistogram which has a pre-defined number
of bins, and will dynamically double the size of those bins as required
to fit all recorded values into a bin.

Changes:
- Implemented `fme.core.histogram.DynamicHistogram` which defines
histograms with a set number of bins of dynamically-increasing size.

- [x] Tests added

---------

Co-authored-by: Brian Henn <brianhenn@allenai.org>