Save histogram data during inference (#374)
This PR adds a new output file `histograms.nc` containing histograms of
prediction and target data. Due to its small size, it is always written.

Data is stored with shape [source, timestep, bin]. If data is not
available for one of the sources (e.g. it is input-only so there is no
prediction data), the histograms will have zero counts.

Changes:
- added a new output file `histograms.nc` containing histograms of
prediction and target data
- made low level logging calls in climate_data.py quieter, as they
already result in warning logs at a higher level due to None being
returned
- added i_time_start argument to `DynamicHistogram.add`

- [x] Tests added