Generalize run_on_batch to ragged time dim. (#364)
This is the first step in implementing inference that can run for more
time steps than are available in the target data, e.g. arbitrarily long
inference.

Changes:
- Refactors the main training loop in `run_on_batch` in order to more
clearly focus on the main operations of the training loop. Helper
functions are created to facilitate this along with simple corresponding
unit tests.
- Sets loss to NaN if target data is missing, e.g. in the case of the
unresolved time-dim of a variable which is only specified at an initial
condition during inference.
- Creates a dummy `ReturnZerosModule` for testing. It has an unused
parameter so that there are parameters to pass to the optimizer but
otherwise, simply makes sure to output the correctly shaped tensor.
- Adds a new check for NaN-valued losses. If we are in training mode and
the loss is NaN, a ValueError is raised which is currently uncaught --
resulting in a crashed training loop.

- [x] Tests added

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>