Remove find_unused_parameters=True (#375)
We are getting the following warning during parallel training:

```
2023-10-12T15:42:17.564450732Z [W reducer.cpp:1267] Warning: find_unused_parameters=True was specified in DDP constructor, but did not find any unused parameters in the forward pass. This flag results in an extra traversal of the autograd graph every iteration,  which can adversely affect performance. If your model indeed never has any unused parameters in the forward pass, consider turning this flag off. Note that this warning may be a false positive if your model has flow control causing later iterations to have unused parameters. (function operator())
```

`find_unused_parameters` is a performance optimization argument to
[DistributedDataParallel](https://pytorch.org/docs/stable/generated/torch.nn.parallel.DistributedDataParallel.html)
which makes the code perform better in handling unused parameters,
because it pre-emptively marks them as ready to be reduced before
computing the outputs of the graph. However, we do not have any unused
parameters so this results in a performance hit instead. Setting this
argument to `False` should not change the output of the model or cause
any bugs, even if unused parameters were to be added.

Changes:
- Set find_unused_parameters to False (default) when initializing
DistributedDataParallel.

- [ ] Tests added