Simplify typing on ClimateData (#400)
Includes some simplifications to the typing of ClimateData which made it
easier to implement #339.

Instead of returning None types to handle missing variables, exceptions
are raised. This makes it much easier to statically type equations using
those variables, instead of checking for None everywhere.

A special `__getattr__` method to get arbitrary data from the underlying
values is only used in tests of that method and never in the training
code, so it was removed. Using it would have required dynamic typing,
which we don't allow in our code.

Changes:
- ClimateData tensor attributes are no longer Optional, instead a
KeyError is raised if they are missing
- `ClimateData.__getattr__` override is removed
- removed `required_names` argument from `derived_variables.register`,
as its purpose now handled automatically by raised KeyErrors if
variables are missing.
- Added a test of `compute_derived_quantities`

- [x] Tests added