Require specifying a full data config for inline inference (#403)
Currently the data config for inline inference inherits most of its
properties (except `n_samples` and `batch_size`) from the validation
data config. This results in coupling between the validation data config
and inline inference data config that can be undesirable (e.g. if
wanting to window the samples used for inline inference/validation by
different amounts).

I implemented this as a breaking change (in training config) but with
more code complexity it could be done in a backwards compatible way.

Changes:
-  new `data` entry in `InlineInferenceConfig`
- removed `n_samples` and `batch_size` items from
`InlineInferenceConfig`

- [x] Tests added

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>