Silence data loader logs during inference (#421)
When using 5-year inference, the logs quickly become large from data
loader initialization log lines. In wandb logs, there is no way to
recover the start of the log when the logs become too long. In beaker,
you have to download a log file to recover the start of the log. The
start of the log is particularly important to be able to trace a wandb
run back to a beaker run, and vice-versa.

This PR resolves the issue by setting the log level to WARNING instead
of INFO during data loading for inference.

Changes:
- Data loader factories during inline and standalone inference now set
log level to WARNING while inference data is being loaded

- [ ] Tests added

Resolves #402