Update to python 3.10 and torch 2.1.2 (#453)
Want to be on more recent versions of torch/python to be able to get
latest features. A downside of this PR is that the more recent NVIDIA
base image is even more bloated (~21GB!). We may want to move away from
using this base image, which has a bunch of stuff we don't need.

Changes:
- use more recent (23.08) NVIDIA pytorch image as our base image
- update various requirements.txt files to use python=3.10 and
torch=2.1.2 in our conda env, roughly consistent with what is in the
NVIDIA docker image
- stop requiring apex! We now use a fused version of Adam that is native
in pytorch, and which apparently has comparable performance since torch
2.0.1. This required a [small
change](https://github.com/ai2cm/modulus/pull/2) to the modulus
repository.

Resolves #450