Inference time coarsening (#446)
This PR allows online time coarsening of prediction outputs during
inference, to reduce data files size for long simulations. When
specified, the inference routine performs time coarsening by an integer
`coarsen_factor` that must be a factor of the `forward_steps_in_memory`
integer parameter. Coarsening is applied to batches of data that are
then sent to various data writers (raw predictions, reduced predictions,
histograms).

Because the first inference memory window is a special case of size
`forward_steps_in_memory` + 1 due to the initial conditions being
included, those initial conditions are not coarsened and are saved as
the first timestep, then coarsening is applied to the rest of the
window. Metrics logged to wandb are not coarsened.

Changes:
- Adds a 'TimeCoarsen' class and associated config class
(`TimeCoarsenConfig`) to the inference config, which coarsens batches of
data prior to passing those batches to the data writers.