Make histogram netCDF optional and configurable (#538)
Computing and writing histograms to disk is expensive.
This PR allows the user to turn off the histogram netCDF. It also makes
the histogram data writer use the `save_names` feature to allow
calculation/saving of only some histograms. It will be required to use
the same set of names for the "raw predictions" as the histograms.

I did change the default to not save the histograms. Since we've found
this to be a bottleneck, I'd prefer the user has to turn it on
intentionally.

- [x] Tests added