Allow subsetting of datasets using a slice (#573)
This makes it possible to use a different period of a single dataset for
training and validation. Useful for AMIP datasets, for example.

Changes:
- replace `DataLoaderParams.window_starts` with
`DataLoaderParams.subset`, since the "window" term is a bit confusing
and overlaps with the "window" concept in our inference code, which is
unrelated to this option.
- deprecate `DataLoaderParams.n_samples` because it is redundant with
`DataLoaderParams.subset.stop`
- for ensemble_xarray case, subset is applied to each member before
concat. So the behavior is changing back to what it was before #563 for
`n_samples`/`subset.stop` case.