Add TrainerConfig.copy_weights_after_batch (#564)
To fine-tune 4-degree models into 1-degree models, we want to be able to
freeze the weights that come from the 4-degree model while still tuning
the new weights pertaining to 1-degree. This means we can't freeze the
layer with requires_grad, as only all or none of a parameter can be
frozen. Instead, we can restore the original model weights in the subset
of the parameter after each batch update.

This PR adds a copy_weights_after_batch option to TrainerConfig to
enable doing this.

Changes:
- Added `get_base_weights` method to `SingleModuleStepperConfig` and
`ExistingStepperConfig`, used by training for
`copy_weights_after_batch`.
- Added CopyWeightsConfig` used as attribute
`TrainerConfig.copy_weights_after_batch`, which is applied after each
epoch if base weights exist.

- [x] Tests added