Add timer for final writer flush (#604)
Currently the final `writer.flush()` is not included in any of the
timers. Also we don't include the writing of the reduced diagnostics
time-mean or zonal-mean diagnostics (although these are probably fast).
This PR updates the timers so that those steps are included in a new
"final writer flush" measurement.