Make saving of checkpoints more verbose (#621)
The logic to determine when to save checkpoints is tricky to get right
and hard to unit test. It is also not obvious to users which epochs the
"best" checkpoints come from. This PR adds a bunch of logging to make
this more explicit.

Changes:
- refactor the checkpoint saving code to a a new
`Trainer.save_all_checkpoints` function
- add logging statements for each type of checkpoint being saved