add L1 loss (#661)
L1 loss is used in the super-resolution literature and this may be a
useful loss for the downscaling project. The existing LpLoss module
unfortunately only supports normalized Lp loss functions, so there is a
need for a simple way to reference pytorch's core L1Loss.

This doesn't resolve #652 but it does make the main change.

Changes:
- Add "L1" as a loss option.