Improved histograms (#651)
Wandb histograms are not well-suited to our needs for a number of
reasons: (1) does not support comparison between runs, (2) does now
allow for log-scaled density (y-axis), (3) does not support multiple
histograms (e.g. target and prediction). This PR switches to using
simple matplotlib histogram plots to facilitate these features.

Changes:
- replaces the native wandb.histogram with our own matplotlib based
plotting
- `DynamicHistogram` is now a comparison metric rather than an intrinsic
metric. This creates a mismatch between the wandb metrics, which plot
target and prediction on a single figures, and the underlying data which
has separate histograms for target and prediction. This mismatch
requires some massaging in the tests.
- adds percentiles to histograms

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>