Remove unused get method from required API (#676)
Currently the aggregator specification requires all aggregators for
downscaling provide a `get` method. This get method is only used in
tests. It could in theory be used to retrieve raw data for some new use,
but we very soon plan to add get_dataset methods for this purpose.

This PR relaxes the restriction on aggregators for downscaling so a get
method is not required. Individual aggregators can continue to have and
use a `get` method, and this can be good to use as an intermediate step
before processing for get_wandb, but it is not a property required for
all aggregators, or required by the _ComparisonAggregator protocol.

Changes:
- Removed unused `get` from `_ComparisonAggregator`
- Removed unused `get` from `ComparedDynamicHistogram`
- Removed unused `get` from `fme.downscaling.aggregators.Aggregator`

- [ ] Tests added