Measure training speed in samples per second (#681)
This adds a new metric "training_samples_per_second". The metric is
equal to `batch_size / duration_of_single_batch` and is averaged over
`config.log_train_every_n_batches` batches and then reported along with
the "batch loss".

Alternatively, we could report a "per GPU" number, but I think this is
an easier one to understand and users can scale by # of GPUs if/when
necessary.