swinir multiple inputs outputs (#689)
We want to start testing which input fields make an impact for
downscaling. However, SwinIR currently hard codes the number of output
channels to be equal to the number input channels (= 3 since they assume
RGB input images). This PR remedies this, by simply changing the
hard-coded output channels to a parameter.

Changes:
- SwinIR now accepts an new argument, `out_chans`

- [x] Tests added