Allow subsetting datasets via timestamps (#777)
It is useful to be able to declare the desired subset of a dataset by
timestamps instead of indices. In particular, for the case of
concatenating datasets that each span different time ranges (e.g. FV3GFS
AMIP and ERA5), this is necessary if we want a consistent set of times
or consistent train/valid split for each dataset.

Changes:
- `DataLoaderConfig.subset` is now of type `Union[Slice, TimeSlice]`
where `TimeSlice` is a new class whose `start_time` and `end_time` are
strings.
- new `as_index_slice` takes a `Union[Slice, TimeSlice]` and
`XarrayDataset` and returns a standard python `slice` with appropriate
start/stop/step.

- [x] Tests added

Resolves part of #773 . Does not address configuring inference initial
conditions via timestamps. Punting on that since it may be separately
addressed by the variety of issues contributing to #468.