Allow not checking metadata/coord sameness for ensemble (#792)
In some cases, want to be able to train on datasets which don't exactly
agree in terms of metadata and/or vertical coordinate. For example,
FV3GFS-AMIP and ERA5.

Changes:
- add `strict_ensemble` option to `DataLoaderConfig` which sets whether
or not sameness of metadata and vertical coord should be enforced. The
option has no effect if `data_type='xarray'`. It defaults to True.