Incorporate fine topo using data requirements for downscaling (#768)
We want to explore the impact of incorporating fine topography on our
downscaling models. This PR adds fine topography as a global attribute,
similar to latitude or area_weights.

Changes:
- Add a boolean configuration option
`DownscalingModelConfig.use_fine_topography` which specifies whether or
not to use fine topography in the model.
- Add `fine_topography` as an optional argument to
`registry.ModuleConfig.build` and implement handling fine_topography for
u-net models where simple concatenation of an additional field suffices.
If specified for SwinIR models, an unimplemented exception is raised.
- `get_topography` function in datasets.py which loads the topography
field from a dataset. Note that this code assumes that topography is
present in every dataset moving forward.

- [x] Tests added