Add PairedRestartWriter wrapper class (#818)
Modify the `RestartWriter` class so that its `append_batch` function
does not take in the unnecessary target data. Add a
`PairedRestartWriter` whose `append_batch` function does take and ignore
the target data. The former case is needed for running inference in the
case where no target data is available.

This is a step towards #468 and is a small TODO I'm carving out from
#781