Delete 'start_sample' arg from append_batch (#821)
Since refactoring the inference to only loop over time (and not
batch/ensemble), we no longer need the `start_sample` arg in the various
`DataWriter.append_batch` functions. This is a follow on refactor PR to
clean this up.

Partially addresses #467 (resolves part related to data writer, but does
not touch aggregators).