Dont require surface temp as forcing for slab ocean (#822)
The `OceanConfig.names` is intended to be used to determine the forcing
variables required by the ocean model (e.g. for upcoming 'no target'
inference). It should not include surface temperature for the slab ocean
case.

This bug was fixed, and the property was renamed to `forcing_names`
(ditto for the corresponding property on `Ocean`) for clarity.