Vendorize diffusion code from EDM for downscaling (#826)
Since we are basically applying EDM to downscaling via the CorrDiff
project, it makes sense to re-use code as much as possible. This is
especially true of tricky numerics which makes up the core diffusion
training and data generation algorithms. This PR vendorizes the
generation and neural network preconditioning code from EDM. This code
will likely need to be modified to handle conditioning on low-resolution
inputs in a later PR.