Add PredictionOnlyDataWriter (#831)
This PR provides a data writer class whose "append_batch" method does
not take any target data as input. This is split off from #781 to make
reviewing easier. It also refactors the `TimeCoarsen` class so that it
can be more easily repurposed for writing data both with and without
target.

Changes:
- add `PredictionOnlyDataWriter`
- rename `TimeCoarsen` to `PairedTimeCoarsen` and add corresponding new
`TimeCoarsen` class which just coarsens a single dict of data.

- [x] Tests added