`UNetDiffusionModule` for conditioning in downscaling diffusion models (#828)
Our existing regression modules maps a coarse input to a fine output.
For diffusion, there are latent variables that represent the state along
the sampling chain as well as a noise noise which is necessary to
condition the neural network. Thus the forward method between regression
and diffusion are fundamentally different. This calls for a new registry
to clearly delineate between these two types of torch modules.

This PR is a first step in this direction. It does not contain the
registry yet but is just the wrapper module that handles conditioning
the latent variable with the coarse input.

Changes:
- `unet_diffusion.py` containing the wrapper class`UNetDiffusionModule`
for handling conditioning p(latent | coarse).

- [x] Tests added