Rename `save_raw_prediction_names` to `names` (#845)
Since this option is used to subset fields for all the data writers, its
name should not be specific to the "raw prediction". I think either
`names` or `save_names` would be good. I went with `names` as the more
compact one—but open to other suggestions. E.g. `variable_names` would
be a bit more specific.