Infer timestep within data loader (#830)
This PR adds an option to infer the timestep of the provided data within
the data loader and use it throughout ACE. In so doing it addresses
#783, since inferring the timestep requires validating that the times
are evenly spaced.

Changes:
- Adds an `infer_timestep` option to the `XarrayDataConfig` and a
`timestep` property to `XarrayDataset`. If `infer_timestep` is `True`
(the default), the timestep is inferred from the raw loaded data. If it
is not possible to infer a timestep from the provided data, e.g. if the
times are not evenly spaced, an error is raised. If `infer_timestep` is
`False` then trying to use the `timestep` property of `XarrayDataset`
will raise an error.
- Threads the inferred timestep through to all the relevant places in
ACE, ensuring that it gets converted to units of seconds wherever that
is necessary.
- The timestep is serialized as an integer number of microseconds when
saving a stepper. This is later used at inference time to validate that
the timestep of the trained model matches the timestep of the reference
data used for inference.

The main files to look at for the timestep inference code are:
-
[`data_loading/_xarray.py`]([fme/fme/core/data_loading/_xarray.py](https://github.com/ai2cm/full-model/pull/830/files#diff-96c2d4d9ca950dd7308e006f4117eda0e1a6cb773d8590190feac81b461b2769))
-
[`data_loading/test__xarray.py`]([fme/fme/core/data_loading/test__xarray.py](https://github.com/ai2cm/full-model/pull/830/files#diff-90d2e726f476c87accf493f8781cabba48c89745fe1ced789ef391250b3eedcf))
-
[`data_loading/config.py`]([fme/fme/core/data_loading/config.py](https://github.com/ai2cm/full-model/pull/830/files#diff-13405ecf923a5f66febca891adfecb02fc0f268a44068ad99c9f84c2ef4dfcdd))
-
[`data_loading/test_data_loader.py`]([fme/fme/core/data_loading/test_data_loader.py](https://github.com/ai2cm/full-model/pull/830/files#diff-761cd5378dd8a6ea5a310e59e452c78a2f54ffe1e23edd7316e274579d1002e0))

- [x] Tests added

Resolves #406
Resolves #783

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>